# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CalendarEvent {
  id: string;
  title: string;
  /** ISO 8601 datetime string */
  start: string;
  /** ISO 8601 datetime string */
  end: string;
  /**
     * Address or venue of the event
     * @nullable
     */
  location?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface GeoLocation {
  lat: number;
  lng: number;
  displayName: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetCalendarEventsParams = {
/**
 * The iCal/Google Calendar URL to fetch
 */
url: string;
/**
 * Optional date filter in YYYY-MM-DD format. If provided, returns only events on that day.
 */
date?: string;
};

export type GeocodeAddressParams = {
/**
 * The address to geocode
 */
address: string;
};