# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEvent {
  id: string;
  title: string;
  /** ISO 8601 datetime string */
  start: string;
  /** ISO 8601 datetime string */
  end: string;
  /**
     * Address or venue of the event
     * @nullable
     */
  location?: string | null;
  /** @nullable */
  description?: string | null;
}